[bookmark] Refactor bookmark api route for consistency
