+++ src/client/services/user.service.ts
@@ -2,6 +2,8 @@ import { ApiResponse } from "@typing/api-response.interface";
 import { IUserPublic } from "@typing/user.interface";
 import axios, { AxiosResponse } from "axios";
 
+const userApiUrl = "/api/user";
+
 /**
  * Gets the user data from the server API.
  *
@@ -11,6 +13,23 @@ import axios, { AxiosResponse } from "axios";
 const getUser = (
   id: string
 ): Promise<AxiosResponse<ApiResponse<{ user: IUserPublic }>>> =>
-  axios.get<ApiResponse<{ user: IUserPublic }>>(`/api/user/${id}`);
-
+  axios.get<ApiResponse<{ user: IUserPublic }>>(`${userApiUrl}/${id}`);
 export { getUser };
+export { toggleBookmark };
+/**
+ * Sends a request to the server API to remove the given lesson from the current user's bookmark (as identified by its
+ * session cookie).
+ *
+ * @param {string} lessonId The lesson to remove
+ * @param {boolean} isAlreadyBookmarked Whether the lesson is already bookmarked
+ * @return {Promise<AxiosResponse<ApiResponse>>} The server response
+ */
+const toggleBookmark = (
+  lessonId: string,
+  isAlreadyBookmarked: boolean
+): Promise<AxiosResponse<ApiResponse>> => {
+  if (isAlreadyBookmarked) {
+    return axios.delete<ApiResponse>(`${userApiUrl}/bookmark/${lessonId}`);
+  }
+  return axios.post<ApiResponse>(`${userApiUrl}/bookmark/${lessonId}`);
+};
